Specify dtype for input tensor in self-attention.py

+++ self-attention.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 n_embd = 10
 block_size = 10
 
-inputV = torch.randn((block_size, n_embd),)
+inputV = torch.randn((block_size, n_embd), dtype=torch.float32)
 
 query = nn.Linear(n_embd, n_embd, bias=False)
 key   = nn.Linear(n_embd, n_embd, bias=False)
